move test_evaluate to the test_utils.metrics package
